Moved methods over to appropriate class
